perf: remove stores
