+++ src/index.ts
@@ -1,6 +1,2 @@
-export * from "./lib/ErelaClient";
-export * from "./lib/entities/Node";
-export * from "./lib/entities/Player";
-export * from "./lib/entities/Queue";
-export * from "./lib/entities/Track";
-export * from "./lib/utils/Utils";
+export { Manager } from './structures/Manager';
+export { Player } from './structures/Player';
